requirements.txt
# # LiveKit & Real-time Communication
# livekit-api==1.1.0
# livekit-agents[openai,silero,turn-detector]==1.4.2
# # livekit-plugins-openai==1.3.12
# # livekit-plugins-silero==1.3.12
# # livekit-plugins-turn-detector==1.4.2
# # livekit-server-sdk==2.14.2

# # FastAPI & Web Framework
# fastapi==0.128.0
# uvicorn==0.40.0
# python-multipart==0.0.20

# # Environment Variables
# python-dotenv==1.2.1

# # Document Processing
# pdfplumber==0.11.0
# python-docx==1.0.0

# # AI & LLM
# openai
# langfuse==3.12.1

# # Streamlit (UI Dashboard)
# streamlit==1.28.1

# # HTTP Requests
# requests==2.31.0


# LiveKit & Real-time Communication
livekit-api==1.1.0
livekit-agents[openai,silero,turn-detector]==1.4.4

# FastAPI & Web Framework
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.20

# Environment Variables
python-dotenv==1.2.1

# Document Processing
pdfplumber==0.11.0
python-docx==1.0.0
langchain-openai==1.1.7
# AI & LLM
openai>=2
langfuse==3.12.1

# HTTP Requests
requests==2.31.0
psycopg2-binary==2.9.9